Update on millionsong.py
Changed the random V generation to numpy.random.choice([-1,1],(rows,cols)) which
will generate a matrix of type (rows,cols) with values either -1 or 1.

Changed in find_exact_cosine_distance to check in the real bucket.
Before it only checked the length of a tuple, which in this case was 2
every time.